test: use int array indices in item-alloc tests (silence recurring CS1503 false positive)
The Copilot bot re-flagged 12 inline threads on PR #833 claiming
`rom.Data[addr + i]` / `bytes[ItemAddr + n]` "won't compile (CS1503)" because
the index is uint. That is a false positive — C# array element access accepts
int/uint/long/ulong indices, and these tests already compiled + passed
(Core.Tests 2224, Avalonia.Tests 7219). To stop the recurring bot noise and let
the threads resolve cleanly, convert every uint array-index expression in the
item-alloc tests to an int index — no behavior change.

- ItemAllocCoreTests.cs: hoist `int a = (int)addr;` in the three template-assert
  blocks and index `rom.Data[a + N]`; change the three `for (uint i ...)` byte
  loops to `for (int i ...)`; cast the helper record writes to
  `bytes[(int)ItemAddr + N]` and the wired-seam assert to `rom.Data[(int)addr + 1]`.
- ItemNewAllocViewModelTests.cs: `bytes[(int)ItemAddr + 6]` in MakeRom.
- ItemNewAllocScreenshotTest.cs: no array-index sites (already only method
  calls like rom.u32(addr+12)), so nothing to change there.

19 index sites converted. Verified: dotnet build Core.Tests + Avalonia.Tests
succeed; Core.Tests ItemAlloc (13) + Avalonia ItemNewAlloc (6) green. Feat
branch stays code-only (no pr-screenshots in the diff).

+++ FEBuilderGBA.Avalonia.Tests/ItemNewAllocViewModelTests.cs
@@ -24,7 +24,7 @@ namespace FEBuilderGBA.Avalonia.Tests
             for (int i = bytes.Length / 2; i < bytes.Length; i++) bytes[i] = 0xFF;
             var rom = new ROM();
             rom.LoadLow("synthetic-fe8u.gba", bytes, "BE8E01");
-            bytes[ItemAddr + 6] = 0x01; // item number (record is plausible)
+            bytes[(int)ItemAddr + 6] = 0x01; // item number (record is plausible)
             rom.LoadLow("synthetic-fe8u.gba", bytes, "BE8E01");
             return rom;
         }
